Add files via upload
In this commit, I have fully finished the Javascript file. I have finished provviding all of the necessary comments, and I believe that my project is relatively easy to follow. Additionally, I finally got the function that random selects an employee, to work. The only problem is that there is currently a lot of hard coding involved, and the max employee count that it works for is 10. Overall, this app is very functional, while there are a few small things that I wold love help fixing, I can say I am happy with how it currently functions.